server: fix handing of a connection closed by peer
When a peer closed a connection (e.g. after we send a notification),
rx goroutine finds it immediately since read() returns an error and
kills tomb. The problem is that tx goroutine doesn't check if tomb
Dying() so tx doesn't die until tx tries to write keepalive to the
socket (and it doesn't never happen if keepalive interval is zero). So
fsm does't become idle shortly.

This fixes tx to check Dying() and makes sure that it sends
notification messages befor dying if they exists.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>