peer: invert the order of pathattr update and adj-rib storing
One of the adj-rib's objective is debugging of raw bgp messages which
BGP speaker received and sent.
From this point of view, it's desirable to alter pathattr after storing to
adj-rib-in when receiving, and before stroring to adj-rib-out when sending.

After this change, we must update as-path path attr to 4byte when we
retrieve paths from adj-rib-in. This is performed in sendPathToSiblings()