tools: avoid adding a new option to bgpyang2golang.py and fix gobgp.yang
pyang has an internal flag "multiple_modules" to handle multiple yang modules.
use it instead of introducing a new option --augment

also, fix warnings and errors in gobgp.yang which were there but not has been
detected for a while due to the usage of --augment.
by using "multiple_modules" flag, pyang detected them.

most notable error is wrong usage of list, which needs a key in its
substatement.

this patch adds key to some of them, and for others uses leaf-list
instead.

Signed-off-by: ISHIDA Wataru <ishida.wataru@lab.ntt.co.jp>