server: fix fsm.keepaliveTicker race
When fsm state goes to idle from established, fsm.keepaliveTicker is
set to nil. This can happen before <-h.t.Dying() case in
sendMessageloop().

This removes fsm.keepaliveTicker. Ticker is created locally. With
this, a keepalive message could be sent from openconfirm to
established with a shorter interval. But it should not break anything.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>