server: fix fixed timeout for write
using 30 secs is a bad idea. Let's use negotiated holdtime
instead. But I think that we should not use timeout for write. when a
timeout for write expires, we shuld just check:

a) if the conneciton is still valid
b) a peer is not de-configured

a) nor b) is not the case, we should try write again.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>