table: fix bug of incomplete withdrawal handling
When mod action is used in route server configuration,
`Destination.knownPathList` holds multiple modified paths whose original path
is same.
In that case, when the original path is withdrawn, we must remove all
modified paths.
There was a wrong `break` in `Destination.explicitWithdraw()` which avoids
this.
test is also added to check the behavior.

Signed-off-by: ISHIDA Wataru <ishida.wataru@lab.ntt.co.jp>