fix community-set match all evaluation
Under the following condition:

[policy-definitions.statements.conditions.bgp-conditions.match-community-set]
  community-set = cs1
  match-set-options = "all"

match evaluation should return true if all the elements of cs1 are
included in the community-set path attr.

For instance, with the aforementioned condition if cs1 = [c0, c1, c2]
and the actual community path attr = [c0], it should be evaluated as
false. However the current code without this patch returns true.