mrt: add an option to only select the best path
In fact, we assume the best path is the first one. There is no such
guarantee but since the main goal is to use less memory, this may work
for some people. Ideally, we should reuse what is done in
destination/path.go but the types are not compatible. Unsure about the
best way forward here. There is an obvious optimisation to not loop over
all entries if we need only the first, but the goal would be sort
correctly and this optimisation wouldn't hold.

The other problem is that we assume that each MRT segment will carry a
new prefix. This is usually true for MRT dumps. I don't think this is
important.