test/lib/gobgp: Use local GoBGP executables in contaier
Currently, to reflect the modification of the local source code or to
switch the base image for GoBGP container, it is required to re-build
the GoBGP container image, and this take a long time to test or debug
with the scenario tests.

This patch fixes to use the local GoBGP executables (gobgp and gobgpd)
in container, and enables to reflect the changes without rebuidling
container image.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>