table: reduce allocs in NewDestination by a factor of 3(ipv4)-5(ipv6)
In NewDestination we add additional path to AddrToRadixkey to check
the concrete type of nlri. If it is one of IP{,v6}AddrPrefix we will
use the existing Prefix and Length to avoid 2 allocs in ParseCidr
and 1 alloc in WriteString. 2 additional allocations are saved for
ipv6 due the larger RadixKey size the buffer ends up growing, since
it also escapes both need heap allocated. I bypass this by creating
a single buffer and passing 8 byte chunks of the slice header to
byteToBase2Buf.