table: reduce allocations in destination selection by a factor of 5
This round of changes is slightly more aggressive than the last but
provides a flat reduction in allocations for the Select paths by an
additional factor of 5.

In table/path.go we provide a new reset method akin to clone, see
inline comments.

In table/destination.go we recalculate the RadixKey we can now derive
our new route destinations with the prior changeset in place to prevent
from the previous, skipping this expensive step. We then reuse the
existing paths at line 1012 that have already been created calling the
newly added reset() method mentioned above.