config: Initialize Graceful Restart State
Currently, the restarting speaker ends restarting procedure
before receiving End-of-Rib markers from all helper speakers.
It is because `MpGracefulRestart.State.Enabled` is not initialized,
always evalueted to false, so the restarting speaker misunderstands
as there is not helper speaker.

This patch adds code to initialize `MpGracefulRestart.State.Enabled`.
This fixes #1296.

Signed-off-by: Satoshi Fujimoto <satoshi.fujimoto7@gmail.com>