flowspec: Set AND to correct operand for TCP Flags
Currently, if TCP Flags field is specified from CLI command,
GoBGP sets 'AND' bit to the invalid operand.
RFC 5575 says it represents the previous term is logically ANDed or not,
and "It should be unset in the first operator byte of a sequence".
This patch fixes this to comply with RFC5575.

Signed-off-by: Satoshi Fujimoto <satoshi.fujimoto7@gmail.com>