config/default: Fix unfilled "neighbors.afi-safis.config.enabled"
Currently, if "neighbors.afi-safis.config.enabled" is omitted, this
field will not initialised as expected when GoBGPd reads config file,
but will initialised when AddNeighbor API is called.
This cause the miss comparison whether reloaded neighbor config is
updated or not (always determined updated).

This patch fixes the condition more strictly and solves this problem.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>