config/default: Avoid setting doubly defaults of neighbor
Currently, the function for setting the default values for neighbors
will called doubly when the neighbor config was added via local config
files (called by "config/serve.go" and AddNeighbor() in
"server/server.go").

This patch enables to detect the duplicated call of this function and
avoid setting the unexpected default values.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>