server: fix fd leak in DialTCP() on Linux
A new socket was created so we must close it before this function
returns _either_ on failure or success. On success, net.FileConn() in
newTCPConn() increases the refcount of the socket so this fi.Close()
doesn't destroy the socket; the caller can use it (and needs to close
it later).

Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>