api: GetPath API alternative to GetRib
With the extremely large RIBs, GetRib gRPC API hits the limits of each
message size, because this API is using simple request/response RPC.

This patch introduces a new API GetPath which using streaming RPC and
can retrieve RIBs even if extremely large.

Note: GetRib can be replaced with GetPath and obsoleted in the future.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>