server: Avoid local paths left on iBGP peers
Currently, with iBGP peering, the CLI injected path will not be withdrawn
when the deletion if any other path exists on RIB, then the CLI injected
path will be left on iBGP peers even if the route is deleted on the
originator router.
This problem can cause routing loops in some case.

This patch fixes to withdraw the old best path when it was the CLI (or
gRPC) injected path.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>