packet/bgp: Avoid Varied Flag Representations in Flowspec Filter
Currently, the parsing results of flags in flowsspec filter
will be varied every time they are parsed.
For example, the tcp-flag '=UFP' may be represented to '=UPF' or '=PUF'
and so on.
This is due to the use of interation to map. Iterations over maps
does not happen in a consistent order.

This fixes it by iterating sorted slices instead of maps.

Signed-off-by: Satoshi Fujimoto <satoshi.fujimoto7@gmail.com>