evpn.md: Update tutorial using BaGPipe
To interchange routes over iBGP session, this patch add the config of
Route Reflector to GoBGP and updates output examples of GoBGP and
BaGPipe.

Note: The BaGPipe Docker image described on this document does not seem
to be maintained for 2 years and contains the old version of BaGPipe. So
this patch drops the description of this image.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>