cli: Missing keywords for EVPN Prefix Route
Currently, the keywords "rt" and "encap" for EVPN Prefix Route are
unexpectedly ignored when parsing the inputs from "gobgp" command.
Also, "label" is not described in the usage.

This patch fixes to pass these keywords to be advertised and updates the
usage format.

Note: This patch makes "rt" as optional field in the usage because "rt"
can be omitted when deleting.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>