cli: Support keyword args for EVPN macadv and multicast
For improving readability of arguments, this patch supports keyword
arguments "etag" and "label" when adding EVPN "macadv" and "multicast"
routes.
Also, this change makes it more easy to introduce new arguments.

With this patch, both of the following formats are supported for the
backward compatibility.

New Format:
 <MACADV>    : <mac address> <ip address> etag <etag> label <label> rd <rd> [rt <rt>...] [encap <encap type>]
 <MULTICAST> : <ip address> etag <etag> rd <rd> [rt <rt>...] [encap <encap type>]

Old Format:
 <MACADV>    : <mac address> <ip address> <etag> <label> rd <rd> [rt <rt>...] [encap <encap type>]
 <MULTICAST> : <ip address> <etag> rd <rd> [rt <rt>...] [encap <encap type>]

Note: This patch makes "rt" as optional field in the usage because "rt"
can be omitted when deleting.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>