packet/bgp: Avoid raw binary in EVPN ESI string
Currently, the value field of EVPN ESI might be corrupted when ESI type
is ARBITRARY (not "single-homed") or unknown type because the value
field will be formatted as the raw byte slice in this case.

This patch fixes to convert the byte slice into the colon separated hex
values which is corresponding to the format of the CLI input.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>