packet/bgp: Use decimal AS number in EVPN ESI
Currently, the representation of AS number in EVPN ESI Type AS (Type 5)
uses the colon separated 2 octet values format, but in other places,
the AS4_PATH attribute for example, the single decimal format is used.

This patch fixes to use the decimal AS number for the consistency with
the AS_PATH/AS4_PATH attributes and also the CLI input format.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>