packet/bgp: Use regexp to parse FlowSpec rules
Currently, the parser for the string representation of FlowSpec rules
splits the given string into characters and validates them character
by character.
So we need to handle the unexpected white spaces carefully.

This patch introduces the regular expressions to parse the FlowSpec
rules and simplify the parsers.
Also improves robustness for the unexpectedly inserted white spaces
like "& == tcp".

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>