cli: Return error when no RD for VPN FlowSpec
Currently, if Route Distinguisher is not specified for the VPN FlowSpec
types via "gobgp" command, the message as following will be displayed:

$ gobgp global rib -a l2vpn-flowspec add match vid 100 then accept
rpc error: code = Unknown desc = attribute type length is short

But this message does not describe which argument should be specified.
This patch fixes to output the message about lack of "rd <rd>".

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>