packet/bgp: Use fixed len types in FlowSpec components
Currently, the "FlowSpecComponentItem" uses int type for the "Op" and
"Value" fields, but Golang int type has 4 bytes length on the 32-bit
env and 8 bytes length on the 64-bit env.
For example, to support the SNAP (Type 20) rule, which has 5 bytes
value field, int type has not enough length on 32-bit env.

This patch fixes to use the fixed length integer types for the
FlowSpec components and avoid overflows of value range.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>