Fix for Zebra configuration via GoBGP API
Currently if user specifies "directly-connected" as an option for route_types in EnableZebraRequest to enable zebra configuration, the RPC request fails with error message:
"unknown route type: directly-connected"

And if I use string "connect" (the one that is used in zebra.zapi.go, it fails with:
"invalid InstallProtocolType: connect"

The reason for this error is, mismatch in string used to identify "connect" type in zapi.go: routeTypeValueMap and bgp_configs.go:InstallProtocolType. I believe the intention of the design is to use the strings as specified in zapi.go. So, changed the string "directly-connected" to "connect" which is a string that is used in zapi.go.