cli: Slash separated offset arg for FlowSpec Prefix
Currently, "gobgp" command supports only;
  "<ipv6 prefix>/<len> <offset>"  # space separated
format to specify the prefix offset for the FlowSpec destination/source
rules, but the displayed format with "rib" command is;
  "<ipv6 prefix>/<len>/<offset>"  # slash separated
and the slash separated format is not allowed as command arguments.

This patch enables to specify the slash separated format for the
FlowSpec destination/source rules arguments and resolve unmatch of the
input/output formats.

Example:
$ gobgp global rib -a ipv6-flowspec add match destination 0:db8:1::1/64/16 then accept
$ gobgp global rib -a ipv6-flowspec
   Network                        Next Hop             AS_PATH              Age        Attrs
*> [destination: 0:db8:1::/64/16] fictitious                                00:00:00   [{Origin: ?}

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>