server: Empty path list in initial WatchEventUpdate
Currently, MonitorRib API with ADJ_IN type and "current" flag returns
the initial response with an empty path list when gRPC client connecting
to GoBGP daemon.
This causes the client can not receive the current paths which GoBGP
daemon already has.

This patch fixes to compose the initial MonitorRib response with the
current paths.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>