fix breakage with the latest github.com/satori/go.uuid
The API has changed with the master branch of go.uuid. GoBGP uses the
dependency management tool so it's not problem. However, there are
projects using GoBGP as a library and doesn't use a dependency
management tool...

GoBGP has used a released version of go.uuid so now I have to change
Gopkg.toml to use tha master branch.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>