test/lib/gobgp: Adding routes without reloading
With the current implementation, when adding routes on GoBGPContainer,
we re-configure and reload daemon, then add routes using CLI command.
Differ from QuaggaBGPContainer, when deletion of routes, reloading is
not required, but if reloading daemon frequently, we need to wait for
the daemon restarting, it causes testing time longer and unstable.

This patch enables to add routes without reloading daemon.

Note: According to this change, adding routes should be called after
starting daemons.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>