travis: Introduce markdownlint
When using IDE (e.g. VSCode), we can easily check the results of
markdownlint and it should not be a heavy burden to write docs.
This patch introduces an unit test to invoke markdownlint on Travis-CI.

See https://github.com/DavidAnson/markdownlint/blob/master/doc/Rules.md,
for configuration details of ".markdownlint.json".

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>