fsm: Don't wait EoR if the peer is restarting
RFC4724(Graceful Restart) says that if "Restart State" bit is set
in the Open message, its peer must not wait for the EoR marker
from the speaker before advertising routing information to the speaker.
But currently, GoBGP waits EoR even if "Restart State" bit is set
in graceful restart process.

This commit fixes it to skipping syncing process for the
restarting peers.

Signed-off-by: Satoshi Fujimoto <satoshi.fujimoto7@gmail.com>