server/server: Broadcast BMPPeerDownMessage for peer deletion via CLI
Currently, GoBGP does not send BMPPeerDownMessage when its peer is deleted
via CLI.

When a peer is deleted, the connection will be lost and FsmMsg will be sent
to BgpServer. Then, BgpServer checks FsmMsg whether the peer in FsmMsg exists
in BgpServer's neighbor map.
In this case, the peer has already deleted from neighborMap, so BgpServer
regards the FsmMsg as invalid and discards it.

This commit fixes it by holding the neighbor in neighborMap until receiving
FsmMsg from the neighbor.

Signed-off-by: Satoshi Fujimoto <satoshi.fujimoto7@gmail.com>