test/lib/quagga: Minimize advertisement-interval
Quagga seems to advertise routes to neighbors by intervals of 30 seconds
(eBGP) or 5 seconds (iBGP) by default. Because of this delay, timer in
"bgp_confederation_test.py", 60 seconds, can be exceeded when the UPDATE
messages need to go through multi ASs.

This patch fixes to use minimum advertisement-interval and reduces the
time of test cases.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>