zclient: Drop NextHop Tracking dampening feature
The dampening feature was implemented to delay the update of nexthop
states in order to avoid the states are "extremely" frequent updated.
But with this implementation, if a path is withdrawn by peer or via API
which the dampening feature delaying the update of that path, the
withdrawn path can be restored unexpectedly again.

And currently Quagga and FRRouting does not support the dampening, this
path drops the dampening feature to avoid this problem.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>