api: Allow multiple NLRIs in single MpReachNLRIAttribute
Currently, api.UnmarshalPathAttributes() does not allow the multiple
NLRIs in a single MpReachNLRIAttribute because when AddPath API supposes
only single NLRI is composed. This assumption is equivalent to that
"message Path" structure should have single NLRI for calling AddPath
API.

But to represent the received UPDATE messages from peers, a single
MpReachNLRIAttribute can contain the multiple NLRIs.

This patch fixes to allow the multiple NLRIs in a single
MpReachNLRIAttribute and removes this constraint.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>