log: don't use global logger from logrus
Instead, use our own global logger. It is defined in config package
and users should be able to configure it this way. Example has been
updated to reflect this change.

Unfortunately, for the user, this is not totally a drop-in change:

 - Get the logger: `log := config.Logger`.
 - Remove import `github.com/sirupsen/logrus`, except when using
 - levels (like `logrus.DebugLevel`). In this case, don't name the
 - import as `log`, but keep `logrus`.
 - If using `log.SetOutput`, use `log.Output` instead.
 - If using `log.SetFormatter`, use `log.Formatter` instead.

With recent versions of logrus, no other changes are needed.

While this change is not as neat as to have a per-server logger, this
should fit most uses where we want to control how GoBGP do the logging
as a library.

See #1638.